+++ lib/acts_as_moonable.rb
@@ -2,6 +2,7 @@ require "acts_as_moonable/version"
 require "acts_as_moonable/railtie"
 require "acts_as_moonable/core_ext"
 require "acts_as_moonable/humanable"
+require "acts_as_moonable/interval"
 
 module ActsAsMoonable
   # Your code goes here...