+++ test/dummy/app/models/application_record.rb
@@ -1,3 +1,6 @@
 class ApplicationRecord < ActiveRecord::Base
+  include ActsAsMoonable::Humanable
+  include ActsAsMoonable::Interval
+
   self.abstract_class = true
 end